Rename uiUtilityQSOData.{h,c} to uiUtilities.{h,c}.
