+++ firmware/source/hardware/fw_HR-C6000.c
@@ -23,7 +23,7 @@
 #include "fw_settings.h"
 #include <SeggerRTT/RTT/SEGGER_RTT.h>
 #include <user_interface/menuHotspot.h>
-#include <user_interface/uiUtilityQSOData.h>
+#include <user_interface/uiUtilities.h>
 #include "fw_trx.h"
