+++ firmware/source/user_interface/menuContactList.c
@@ -17,7 +17,7 @@
  */
 #include <fw_codeplug.h>
 #include <user_interface/menuSystem.h>
-#include <user_interface/uiUtilityQSOData.h>
+#include <user_interface/uiUtilities.h>
 #include <user_interface/uiLocalisation.h>
 #include "fw_main.h"
 #include "fw_settings.h"
