+++ firmware/source/user_interface/uiUtilities.c
@@ -20,7 +20,7 @@
 #include <hardware/fw_HR-C6000.h>
 #include <hardware/fw_SPI_Flash.h>
 #include <user_interface/menuSystem.h>
-#include <user_interface/uiUtilityQSOData.h>
+#include <user_interface/uiUtilities.h>
 #include <user_interface/uiLocalisation.h>
 #include "fw_trx.h"
 #include "fw_settings.h"
